fixes for LEAP-872, LEAP-935, LEAP-942, including moving the auto-accept checkbox to the bottom bar, fixing reject all button not working, and removing the regions use for interactivity
