+++ web/libs/editor/src/components/AnnotationTab/DynamicPreannotationsToggle.js
@@ -1,30 +1,13 @@
 import { inject, observer } from 'mobx-react';
 import { useEffect } from 'react';
-import { IconCheck, IconCross } from '../../assets/icons';
-import { Button } from '../../common/Button/Button';
 import { Space } from '../../common/Space/Space';
 import Toggle from '../../common/Toggle/Toggle';
 import ToolsManager from '../../tools/Manager';
 import { Block, Elem } from '../../utils/bem';
 import './DynamicPreannotationsToggle.styl';
 
-const injector = inject(({ store }) => {
-  const annotation = store.annotationStore?.selected;
-  const suggestions = annotation?.suggestions;
 
-  return {
-    store,
-    annotation,
-    suggestions,
-    interfaces: Array.from(store?.interfaces),
-  };
-});
-
-export const DynamicPreannotationsToggle = injector(observer(({
-  store,
-  annotation,
-  suggestions,
-}) => {
+export const DynamicPreannotationsToggle = inject('store')(observer(({ store }) => {
   const enabled = store.hasInterface('auto-annotation') && !store.forceAutoAnnotation;
 
   useEffect(() => {
@@ -49,16 +32,7 @@ export const DynamicPreannotationsToggle = injector(observer(({
             label="Auto-Annotation"
             style={{ color: '#7F64FF' }}
           />
-          {suggestions.size > 0 && (
-            <Space size="small">
-              <Elem name="action" tag={Button} mod={{ type: 'reject' }} onClick={() => annotation.rejectAllSuggestions()}>
-                <IconCross/>
-              </Elem>
-              <Elem name="action" tag={Button} mod={{ type: 'accept' }} onClick={() => annotation.acceptAllSuggestions()}>
-                <IconCheck/>
-              </Elem>
-            </Space>
-          )}
+          
         </Space>
       </Elem>
     </Block>
